Several changes:
- Add static build option.
- Rearrange source and include files to satisfy CMake's dislike
  for top-level include files.
- Separate C and C++ code in both source files and headers.
