+++ compile_flags.txt
@@ -1,0 +1,5 @@
+-I/Users/46774035/src/tci/include
+-I/Users/46774035/build/tci/include
+-I/opt/homebrew/include
+-xc++
+-std=c++20
