+++ include/tci.hpp
@@ -1,0 +1,12 @@
+#ifndef _TCI_HPP_
+#define _TCI_HPP_
+
+#include "tci/barrier.hpp"
+#include "tci/communicator.hpp"
+#include "tci/context.hpp"
+#include "tci/mutex.hpp"
+#include "tci/parallel.hpp"
+#include "tci/slot.hpp"
+#include "tci/work_item.hpp"
+
+#endif
