+++ include/tci/mutex.hpp
@@ -1,55 +1,7 @@
-#ifndef _TCI_MUTEX_H_
-#define _TCI_MUTEX_H_
+#ifndef _TCI_MUTEX_HPP_
+#define _TCI_MUTEX_HPP_
 
-#include "tci_global.h"
-
-#include "yield.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#if TCI_USE_OS_UNFAIR_LOICK
-
-typedef os_unfair_lock tci_mutex;
-
-#elif TCI_USE_OSX_SPINLOCK
-
-typedef OSSpinLock tci_mutex;
-
-#elif TCI_USE_PTHREAD_SPINLOCK
-
-typedef pthread_spinlock_t tci_mutex;
-
-#elif TCI_USE_OMP_LOCK
-
-typedef omp_lock_t tci_mutex;
-
-#elif TCI_USE_PTHREAD_MUTEX
-
-typedef pthread_mutex_t tci_mutex;
-
-#else
-
-typedef char tci_mutex;
-
-#endif
-
-int tci_mutex_init(tci_mutex* mutex);
-
-int tci_mutex_destroy(tci_mutex* mutex);
-
-int tci_mutex_lock(tci_mutex* mutex);
-
-int tci_mutexrylock(tci_mutex* mutex);
-
-int tci_mutex_unlock(tci_mutex* mutex);
-
-#ifdef __cplusplus
-}
-#endif
-
-#if defined(__cplusplus) && !defined(TCI_DONT_USE_CXX11)
+#include "tci/mutex.h"
 
 #include <system_error>
 
@@ -108,7 +60,5 @@ class mutex
 };
 
 }
-
-#endif
 
 #endif
