+++ include/tci/pipeline.hpp
@@ -1,50 +1,7 @@
-#ifndef _TCI_PIPELINE_H_
-#define _TCI_PIPELINE_H_
+#ifndef _TCI_PIPELINE_HPP_
+#define _TCI_PIPELINE_HPP_
 
-#include "tci_global.h"
-
-#include "yield.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-enum
-{
-    TCI_NOT_FILLED,
-    TCI_DRAINING,
-    TCI_FILLING,
-    TCI_FILLED
-};
-
-typedef struct tci_pipeline
-{
-    void* buffer;
-    size_t size;
-    unsigned depth;
-    unsigned last_drained;
-    unsigned last_filled;
-    volatile int state[1];
-} tci_pipeline;
-
-void tci_pipeline_init(tci_pipeline** p, unsigned depth, size_t size, void* buffer);
-
-void tci_pipeline_destroy(tci_pipeline* p);
-
-void* tci_pipeline_drain(tci_pipeline* p);
-
-int tci_pipeline_trydrain(tci_pipeline* p, void** buffer);
-
-void tci_pipeline_drained(tci_pipeline* p, void* buffer);
-
-void* tci_pipeline_fill(tci_pipeline* p);
-
-int tci_pipeline_tryfill(tci_pipeline* p, void** buffer);
-
-void tci_pipeline_filled(tci_pipeline* p, void* buffer);
-
-#ifdef __cplusplus
-}
+#include "tci/pipeline.h"
 
 namespace tci
 {
@@ -66,7 +23,7 @@ class pipeline_base : private Alloc
 
         ~pipeline_base()
         {
-            deallocate(_p.buffer, _p.depth*_p.size/sizeof(T));
+            deallocate(_p->buffer, _p->depth*_p->size/sizeof(T));
             tci_pipeline_destroy(_p);
         }
 
@@ -244,7 +201,5 @@ class pipeline : public pipeline_base<T, Alloc>
 };
 
 }
-
-#endif
 
 #endif
