+++ src/tci/communicator.c
@@ -1,5 +1,5 @@
-#include "communicator.h"
-#include "parallel.h"
+#include "tci/communicator.h"
+#include "tci/parallel.h"
 
 #include <stdlib.h>
 #include <string.h>
@@ -166,8 +166,8 @@ int tci_comm_gang(tci_comm* parent, tci_comm* child,
 
 #if TCI_USE_OPENMP_THREADS || TCI_USE_PTHREADS_THREADS || TCI_USE_WINDOWS_THREADS
 
-static void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
-                           tci_range range, tci_range_func func, void* payload)
+void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
+                    tci_range range, tci_range_func func, void* payload)
 {
     if (n == 1)
     {
@@ -184,9 +184,9 @@ static void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
     func(comm, first*range.grain, TCI_MIN(last*range.grain, range.size), payload);
 }
 
-static void tci_distribute_2d(unsigned num, unsigned idx, tci_comm* comm,
-                              tci_range range_m, tci_range range_n,
-                              tci_range_2d_func func, void* payload)
+void tci_distribute_2d(unsigned num, unsigned idx, tci_comm* comm,
+                       tci_range range_m, tci_range range_n,
+                       tci_range_2d_func func, void* payload)
 {
     if (num == 1)
     {
@@ -214,82 +214,10 @@ static void tci_distribute_2d(unsigned num, unsigned idx, tci_comm* comm,
                nfirst*range_n.grain, TCI_MIN(nlast*range_n.grain, range_n.size), payload);
 }
 
-#elif TCI_USE_TBB_THREADS
-
-static void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
-                           tci_range range, tci_range_func func, void* payload)
-{
-    if (n == 1)
-    {
-        func(comm, 0, range.size, payload);
-        return;
-    }
-
-    range.grain = TCI_MAX(range.grain, 1);
-    uint64_t ngrain = (range.size+range.grain-1)/range.grain;
-
-    tbb::task_group tg;
-
-    for (unsigned idx = 0;idx < n;idx++)
-    {
-        tg.run(
-        [&,idx]
-        {
-            uint64_t first = (idx*ngrain)/n;
-            uint64_t last = ((idx+1)*ngrain)/n;
-
-            func(comm, first*range.grain, TCI_MIN(last*range.grain, range.size), payload);
-        });
-    }
-
-    tg.wait();
-}
-
-static void tci_distribute_2d(unsigned num, unsigned idx, tci_comm* comm,
-                              tci_range range_m, tci_range range_n,
-                              tci_range_2d_func func, void* payload)
-{
-    if (num == 1)
-    {
-        func(comm, 0, range_m.size, 0, range_n.size, payload);
-        return;
-    }
-
-    unsigned m, n;
-    tci_partition_2x2(num, range_m.size, num, range_n.size, num, &m, &n);
-
-    range_m.grain = TCI_MAX(range_m.grain, 1);
-    range_n.grain = TCI_MAX(range_n.grain, 1);
-    uint64_t mgrain = (range_m.size+range_m.grain-1)/range_m.grain;
-    uint64_t ngrain = (range_n.size+range_n.grain-1)/range_n.grain;
-
-    tbb::task_group tg;
-
-    for (unsigned idx_m = 0;idx_m < m;idx_m++)
-    {
-        for (unsigned idx_n = 0;idx_n < n;idx_n++)
-        {
-            tg.run(
-            [&,idx_m,idx_n]
-            {
-                uint64_t mfirst = (idx_m*mgrain)/m;
-                uint64_t nfirst = (idx_n*ngrain)/n;
-                uint64_t mlast = ((idx_m+1)*mgrain)/m;
-                uint64_t nlast = ((idx_n+1)*ngrain)/n;
-
-                func(comm, mfirst*range_m.grain, TCI_MIN(mlast*range_m.grain, range_m.size),
-                           nfirst*range_n.grain, TCI_MIN(nlast*range_n.grain, range_n.size), payload);
-            });
-        }
-    }
-
-    tg.wait();
-}
-
 #elif TCI_USE_OMPTASK_THREADS
 
-static void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
-                           tci_range range, tci_range_func func, void* payload)
+void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
+                    tci_range range, tci_range_func func, void* payload)
 {
     (void)idx;
 
@@ -317,9 +245,9 @@ static void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
     }
 }
 
-static void tci_distribute_2d(unsigned num, unsigned idx, tci_comm* comm,
-                              tci_range range_m, tci_range range_n,
-                              tci_range_2d_func func, void* payload)
+void tci_distribute_2d(unsigned num, unsigned idx, tci_comm* comm,
+                       tci_range range_m, tci_range range_n,
+                       tci_range_2d_func func, void* payload)
 {
     (void)idx;
 
@@ -359,6 +287,8 @@ static void tci_distribute_2d(unsigned num, unsigned idx, tci_comm* comm,
 }
 
 #elif TCI_USE_DISPATCH_THREADS
+
+#include <dispatch/dispatch.h>
 
 typedef struct tci_distribute_func_data
 {
@@ -410,8 +340,8 @@ static void tci_distribute_2d_func(void* data_, size_t idx)
                TCI_MIN(nlast*data->range_n->grain, data->range_n->size), data->payload);
 }
 
-static void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
-                           tci_range range, tci_range_func func, void* payload)
+void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
+                    tci_range range, tci_range_func func, void* payload)
 {
     (void)idx;
 
@@ -430,9 +360,9 @@ static void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
     dispatch_apply_f(n, queue, &data, tci_distribute_func);
 }
 
-static void tci_distribute_2d(unsigned num, unsigned idx, tci_comm* comm,
-                              tci_range range_m, tci_range range_n,
-                              tci_range_2d_func func, void* payload)
+void tci_distribute_2d(unsigned num, unsigned idx, tci_comm* comm,
+                       tci_range range_m, tci_range range_n,
+                       tci_range_2d_func func, void* payload)
 {
     (void)idx;
 
@@ -455,84 +385,23 @@ static void tci_distribute_2d(unsigned num, unsigned idx, tci_comm* comm,
     dispatch_apply_f(m*n, queue, &data, tci_distribute_2d_func);
 }
 
-#elif TCI_USE_PPL_THREADS
+#elif TCI_USE_TBB_THREADS || TCI_USE_PPL_THREADS
 
-static void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
-                           tci_range range, tci_range_func func, void* payload)
-{
-    (void)idx;
-
-    if (n == 1)
-    {
-        func(comm, 0, range.size, payload);
-        return;
-    }
-
-    range.grain = TCI_MAX(range.grain, 1);
-    uint64_t ngrain = (range.size+range.grain-1)/range.grain;
-
-    concurrency::parallel_for(uint64_t(), n,
-    [&](uint64_t idx)
-    {
-        uint64_t first = (idx*ngrain)/n;
-        uint64_t last = ((idx+1)*ngrain)/n;
-
-        func(comm, first*range.grain,
-             TCI_MIN(last*range.grain, range.size), payload);
-    });
-}
-
-static void tci_distribute_2d(unsigned num, unsigned idx, tci_comm* comm,
-                              tci_range range_m, tci_range range_n,
-                              tci_range_2d_func func, void* payload)
-{
-    (void)idx;
-
-    if (num == 1)
-    {
-        func(comm, 0, range_m.size, 0, range_n.size, payload);
-        return;
-    }
-
-    unsigned m, n;
-    tci_partition_2x2(num, range_m.size, num, range_n.size, num, &m, &n);
-
-    range_m.grain = TCI_MAX(range_m.grain, 1);
-    range_n.grain = TCI_MAX(range_n.grain, 1);
-    uint64_t mgrain = (range_m.size+range_m.grain-1)/range_m.grain;
-    uint64_t ngrain = (range_n.size+range_n.grain-1)/range_n.grain;
-
-    concurrency::parallel_for(uint64_t(), m*n,
-    [&](uint64_t idx)
-    {
-        unsigned idx_m = idx % m;
-        unsigned idx_n = idx / m;
-
-        uint64_t mfirst = (idx_m*mgrain)/m;
-        uint64_t nfirst = (idx_n*ngrain)/n;
-        uint64_t mlast = ((idx_m+1)*mgrain)/m;
-        uint64_t nlast = ((idx_n+1)*ngrain)/n;
-
-        func(comm, mfirst*range_m.grain,
-                   TCI_MIN(mlast*range_m.grain, range_m.size),
-                   nfirst*range_n.grain,
-                   TCI_MIN(nlast*range_n.grain, range_n.size), payload);
-    });
-}
+// defined in communicator.cpp
 
 #else // single threaded
 
-static void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
-                           tci_range range, tci_range_func func, void* payload)
+void tci_distribute(unsigned n, unsigned idx, tci_comm* comm,
+                    tci_range range, tci_range_func func, void* payload)
 {
     (void)n;
     (void)idx;
     func(comm, 0, range.size, payload);
 }
 
-static void tci_distribute_2d(unsigned n, unsigned idx, tci_comm* comm,
-                              tci_range range_m, tci_range range_n,
-                              tci_range_2d_func func, void* payload)
+void tci_distribute_2d(unsigned n, unsigned idx, tci_comm* comm,
+                       tci_range range_m, tci_range range_n,
+                       tci_range_2d_func func, void* payload)
 {
     (void)n;
     (void)idx;
