+++ src/tci/context.c
@@ -1,4 +1,4 @@
-#include "context.h"
+#include "tci/context.h"
 
 #include <stdlib.h>
 #include <errno.h>
