+++ src/tci/task_set.c
@@ -1,9 +1,5 @@
-#include "communicator.h"
-#include "task_set.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
+#include "tci/communicator.h"
+#include "tci/task_set.h"
 
 #if TCI_USE_OPENMP_THREADS || TCI_USE_PTHREADS_THREADS || TCI_USE_WINDOWS_THREADS
 
@@ -48,45 +44,6 @@ int tci_task_set_visit(tci_task_set* set, tci_task_func func, unsigned task,
     return 0;
 }
 
-#elif TCI_USE_TBB_THREADS
-
-void tci_task_set_init(tci_task_set* set, tci_comm* comm, unsigned ntask, uint64_t work)
-{
-    set->comm = (tci_comm*)new tbb::task_group();
-    set->ntask = ntask;
-    set->slots = new tci_slot[ntask];
-    for (unsigned task = 0;task < ntask;task++)
-        tci_slot_init(set->slots+task, 0);
-
-    unsigned nt = comm->nthread;
-    unsigned nt_outer, nt_inner;
-    tci_partition_2x2(nt, work, (work == 0 ? 1 : nt),
-                      ntask, ntask, &nt_inner, &nt_outer);
-    tci_comm_gang(comm, &set->subcomm, TCI_EVENLY, nt_outer, 0);
-}
-
-void tci_task_set_destroy(tci_task_set* set)
-{
-    ((tbb::task_group*)set->comm)->wait();
-    delete[] set->slots;
-    delete (tbb::task_group*)set->comm;
-}
-
-int tci_task_set_visit(tci_task_set* set, tci_task_func func, unsigned task,
-                       void* payload)
-{
-    if (task > set->ntask) return EINVAL;
-    if (!tci_slot_try_fill(set->slots+task, 0, 1)) return EALREADY;
-
-    ((tbb::task_group*)set->comm)->run(
-    [set,func,task,payload]
-    {
-        func(&set->subcomm, task, payload);
-    });
-
-    return 0;
-}
-
 #elif TCI_USE_OMPTASK_THREADS
 
 void tci_task_set_init(tci_task_set* set, tci_comm* comm, unsigned ntask, uint64_t work)
@@ -121,6 +78,8 @@ int tci_task_set_visit(tci_task_set* set, tci_task_func func, unsigned task,
 }
 
 #elif TCI_USE_DISPATCH_THREADS
+
+#include <dispatch/dispatch.h>
 
 void tci_task_set_init(tci_task_set* set, tci_comm* comm, unsigned ntask, uint64_t work)
 {
@@ -176,41 +135,9 @@ int tci_task_set_visit(tci_task_set* set, tci_task_func func, unsigned task,
     return 0;
 }
 
-#elif TCI_USE_PPL_THREADS
+#elif TCI_USE_TBB_THREADS || TCI_USE_PPL_THREADS
 
-void tci_task_set_init(tci_task_set* set, tci_comm* comm, unsigned ntask, uint64_t work)
-{
-    (void)comm;
-    (void)work;
-
-    set->comm = (tci_comm*)new concurrency::task_group();
-    set->ntask = ntask;
-    set->slots = new tci_slot[ntask];
-    for (unsigned task = 0;task < ntask;task++)
-        tci_slot_init(set->slots+task, 0);
-}
-
-void tci_task_set_destroy(tci_task_set* set)
-{
-    ((concurrency::task_group*)set->comm)->wait();
-    delete[] set->slots;
-    delete (concurrency::task_group*)set->comm;
-}
-
-int tci_task_set_visit(tci_task_set* set, tci_task_func func, unsigned task,
-                       void* payload)
-{
-    if (task > set->ntask) return EINVAL;
-    if (!tci_slot_try_fill(set->slots+task, 0, 1)) return EALREADY;
-
-    ((concurrency::task_group*)set->comm)->run(
-    [&,func,task,payload]
-    {
-        func(tci_single, task, payload);
-    });
-
-    return 0;
-}
+// defined in task_set.cpp
 
 #else // single threaded
 
@@ -253,7 +180,3 @@ int tci_task_set_visit_all(tci_task_set* set, tci_task_func func,
 
     return ret;
 }
-
-#ifdef __cplusplus
-}
-#endif
