+++ tci.pc.in
@@ -8,3 +8,5 @@ Description: Thread Communicator Interface
 Version: @TCI_VERSION@
 Cflags: -I${includedir}
 Libs: -L${libdir} -ltci
+Requires.private: @PKGCONFIG_REQUIRES@
+Libs.private: @PKGCONFIG_LIBS@
